+++ plugins/clipboard/lang/cs.js
@@ -11,5 +11,6 @@ CKEDITOR.plugins.setLang( 'clipboard', 'cs', {
 	pasteNotification: 'Stiskněte %1 pro vložení. Váš prohlížeč nepodporuje vkládání pomocí tlačítka na panelu nástrojů nebo volby kontextového menu.',
 	pasteArea: 'Oblast vkládání',
 	pasteMsg: 'Vložte svůj obsah do oblasti níže a stiskněte OK.',
-	fileFormatNotSupportedNotification: 'Tento formát souboru není podporovaný. Můžete to zkusit s jedním z podporovaných formátů: ${formats}.'
+	fileFormatNotSupportedNotification: 'Files in ${formats} formats are not supported.', // MISSING
+	fileWithoutFormatNotSupportedNotification: 'File is not supported.' // MISSING
 } );