+++ plugins/clipboard/lang/ru.js
@@ -11,5 +11,6 @@ CKEDITOR.plugins.setLang( 'clipboard', 'ru', {
 	pasteNotification: 'Для вставки нажмите %1. Ваш браузер не поддерживает возможность вставки через панель инструментов или контекстное меню',
 	pasteArea: 'Область вставки',
 	pasteMsg: 'Вставьте контент в эту область и нажмите OK',
-	fileFormatNotSupportedNotification: 'This file format is not supported. You can try with one of the supported formats: ${formats}.' // MISSING
+	fileFormatNotSupportedNotification: 'Files in ${formats} formats are not supported.', // MISSING
+	fileWithoutFormatNotSupportedNotification: 'File is not supported.' // MISSING
 } );