+++ plugins/clipboard/lang/sr.js
@@ -11,5 +11,6 @@ CKEDITOR.plugins.setLang( 'clipboard', 'sr', {
 	pasteNotification: 'Притисните тастер %1 за лепљење. Ваш ретраживач не дозвољаба лепљење из алатне траке или мениа.',
 	pasteArea: 'Залепи зону',
 	pasteMsg: 'Налепите садржај у следећи простор и притисните тастер OK.',
-	fileFormatNotSupportedNotification: 'This file format is not supported. You can try with one of the supported formats: ${formats}.' // MISSING
+	fileFormatNotSupportedNotification: 'Files in ${formats} formats are not supported.', // MISSING
+	fileWithoutFormatNotSupportedNotification: 'File is not supported.' // MISSING
 } );