+++ plugins/clipboard/plugin.js
@@ -124,6 +124,10 @@
 		// jscs:enable maximumLineLength
 		icons: 'copy,copy-rtl,cut,cut-rtl,paste,paste-rtl', // %REMOVE_LINE_CORE%
 		hidpi: true, // %REMOVE_LINE_CORE%
+
+		// File matchers registered by CKEDITOR.plugins.clipboard#addFileMatcher method.
+		_supportedFileMatchers: [],
+
 		init: function( editor ) {
 			var filterType,
 				filtersFactory = filtersFactoryFactory( editor );
@@ -146,75 +150,126 @@
 
 			CKEDITOR.dialog.add( 'paste', CKEDITOR.getUrl( this.path + 'dialogs/paste.js' ) );
 
-			// Convert image file (if present) to base64 string for modern browsers except IE<10, as it does not support
+			// Handle file paste for modern browsers except IE<10, as it does not support
 			// custom MIME types in clipboard (#4612).
-			// Do it as the first step as the conversion is asynchronous and should hold all further paste processing.
-			var isImagePasteSupported = CKEDITOR.plugins.clipboard.isCustomDataTypesSupported || CKEDITOR.plugins.clipboard.isFileApiSupported;
-			if ( isImagePasteSupported && editor.config.clipboard_handleImages ) {
-				var supportedImageTypes = [ 'image/png', 'image/jpeg', 'image/gif' ],
-					unsupportedTypeMsg = createNotificationMessage( supportedImageTypes ),
-					latestId;
+			var isFilePasteSupported = CKEDITOR.plugins.clipboard.isCustomDataTypesSupported || CKEDITOR.plugins.clipboard.isFileApiSupported,
+				latestId;
 
-				editor.on( 'paste', function( evt ) {
-					var dataObj = evt.data,
-						data = dataObj.dataValue,
-						dataTransfer = dataObj.dataTransfer;
+			// Display notification for unsupported file types (#5095).
+			CKEDITOR.plugins.clipboard.addFileMatcher( editor, testImageBase64Support );
+			editor.on( 'paste', function( evt ) {
+				if ( !isFilePasteSupported ) {
+					return;
+				}
 
-					// If data empty check for image content inside data transfer. https://dev.ckeditor.com/ticket/16705
-					// Allow both dragging and dropping and pasting images as base64 (#4681).
-					if ( !data && isFileData( evt, dataTransfer ) ) {
-						var file = dataTransfer.getFile( 0 );
+				var dataObj = evt.data,
+					data = dataObj.dataValue,
+					dataTransfer = dataObj.dataTransfer;
 
-						if ( CKEDITOR.tools.indexOf( supportedImageTypes, file.type ) === -1 ) {
-							editor.showNotification( unsupportedTypeMsg, 'info', editor.config.clipboard_notificationDuration );
+				// Only handle files if there is no data value provided that should take precedence over files.
+				if ( data ) {
+					return;
+				}
 
-							return;
-						}
+				var unsupportedFileTypes = [];
+				for ( var i = 0; i < dataTransfer.getFilesCount(); i++ ) {
+					var file = dataTransfer.getFile( i );
 
-						var fileReader = new FileReader();
-
-						// Convert image file to img tag with base64 image.
-						fileReader.addEventListener( 'load', function() {
-							evt.data.dataValue = '<img src="' + fileReader.result + '" />';
-							editor.fire( 'paste', evt.data );
-						}, false );
-
-						// Proceed with normal flow if reading file was aborted.
-						fileReader.addEventListener( 'abort', function() {
-							// (#4681)
-							setCustomIEEventAttribute( evt );
-							editor.fire( 'paste', evt.data );
-						}, false );
-
-						// Proceed with normal flow if reading file failed.
-						fileReader.addEventListener( 'error', function() {
-							// (#4681)
-							setCustomIEEventAttribute( evt );
-							editor.fire( 'paste', evt.data );
-						}, false );
-
-						fileReader.readAsDataURL( file );
-
-						latestId = dataObj.dataTransfer.id;
-
-						evt.stop();
+					if ( !isFileTypeSupported( file ) ) {
+						unsupportedFileTypes.push( file.type );
 					}
-				}, null, null, 1 );
+				}
+
+				displayUnsupportedFileTypesNotification( unsupportedFileTypes );
+			}, null, null, 1 );
+
+			// Convert image file (if present) to base64 string.
+			// Do it as the first step as the conversion is asynchronous and should hold all further paste processing.
+			editor.on( 'paste', function( evt ) {
+				if ( !isFilePasteSupported || !editor.config.clipboard_handleImages ) {
+					return;
+				}
+
+				var dataObj = evt.data,
+					data = dataObj.dataValue,
+					dataTransfer = dataObj.dataTransfer;
+
+				// If data empty check for image content inside data transfer. https://dev.ckeditor.com/ticket/16705
+				// Allow both dragging and dropping and pasting images as base64 (#4681).
+				if ( data || !isFileData( evt, dataTransfer ) ) {
+					return;
+				}
+
+				var file = dataTransfer.getFile( 0 );
+
+				if ( !testImageBase64Support( file ) ) {
+					return;
+				}
+
+				var fileReader = new FileReader();
+
+				// Convert image file to img tag with base64 image.
+				fileReader.addEventListener( 'load', function() {
+					evt.data.dataValue = '<img src="' + fileReader.result + '" />';
+					editor.fire( 'paste', evt.data );
+				}, false );
+
+				// Proceed with normal flow if reading file was aborted.
+				fileReader.addEventListener( 'abort', function() {
+					// (#4681)
+					setCustomIEEventAttribute( evt );
+					editor.fire( 'paste', evt.data );
+				}, false );
+
+				// Proceed with normal flow if reading file failed.
+				fileReader.addEventListener( 'error', function() {
+					// (#4681)
+					setCustomIEEventAttribute( evt );
+					editor.fire( 'paste', evt.data );
+				}, false );
+
+				fileReader.readAsDataURL( file );
+
+				latestId = dataObj.dataTransfer.id;
+
+				evt.stop();
+			}, null, null, 1 );
+
+			function testImageBase64Support( file ) {
+				var supportedImageTypes = [ 'image/png', 'image/jpeg', 'image/gif' ];
+				return CKEDITOR.tools.indexOf( supportedImageTypes, file.type ) !== -1;
 			}
 
-			// Prepare content for unsupported image type notification (#4750).
-			function createNotificationMessage( imageTypes ) {
-				var humanReadableImageTypes = CKEDITOR.tools.array.map( imageTypes, function( imageType ) {
-					var splittedMimeType = imageType.split( '/' ),
-						imageFormat = splittedMimeType[ 1 ].toUpperCase();
+			function isFileTypeSupported( file ) {
+				var clipboardMatchers = editor.plugins.clipboard._supportedFileMatchers;
+				return CKEDITOR.tools.array.some( clipboardMatchers, function( matcher ) {
+					return matcher( file );
+				} );
+			}
 
-					return imageFormat;
-				} ).join( ', ' ),
+			function displayUnsupportedFileTypesNotification( fileTypes ) {
+				if ( !fileTypes.length ) {
+					return;
+				}
 
-				message = editor.lang.clipboard.fileFormatNotSupportedNotification.
-					replace( /\${formats\}/g, humanReadableImageTypes );
+				fileTypes = CKEDITOR.tools.array.unique( fileTypes );
+				// Make sure to remove unknown file types.
+				fileTypes = CKEDITOR.tools.array.filter( fileTypes, function( fileType ) {
+					return !!CKEDITOR.tools.trim( fileType );
+				} );
 
-				return message;
+				var unsupportedTypeMsg = createNotificationMessage( fileTypes.join( ', ' ) );
+				editor.showNotification( unsupportedTypeMsg, 'info', editor.config.clipboard_notificationDuration );
+			}
+
+			// Prepare content for unsupported file extension notification (#4750).
+			function createNotificationMessage( fileType ) {
+				if ( !fileType ) {
+					return editor.lang.clipboard.fileWithoutFormatNotSupportedNotification;
+				}
+
+				return editor.lang.clipboard.fileFormatNotSupportedNotification.
+					replace( /\${formats\}/g, '<em>' + fileType + '</em>' );
 			}
 
 			// Only dataTransfer objects containing only file should be considered
@@ -1685,6 +1740,30 @@
 	 * @class CKEDITOR.plugins.clipboard
 	 */
 	CKEDITOR.plugins.clipboard = {
+		/**
+		 * Adds a file matcher verifying if a file should be supported via clipboard operations.
+		 *
+		 * In the case of pasting or dragging and dropping unsupported files,
+		 * the clipboard plugin will show a notification informing a user that a given file type is not supported.
+		 *
+		 * ```javascript
+		 * CKEDITOR.plugins.clipboard.addFileMatcher( editor, function( file ) {
+		 * 	var supportedImageTypes = [ 'image/png', 'image/jpeg', 'image/gif' ];
+		 * 	return CKEDITOR.tools.indexOf( supportedImageTypes, file.type ) !== -1;
+		 * } );
+		 * ```
+		 *
+		 * **Note:** This feature won't cancel the `paste` event in the case of an unsupported file.
+		 * It's the integrator's responsibility to properly handle incorrect files
+		 * (e.g. by verifying if the file should be indeed uploaded via upload integrations).
+		 *
+		 * @since 4.19.0
+		 * @param {CKEDITOR.editor} editor The editor instance.
+		 * @param {Function} matcher File matcher.
+		 */
+		addFileMatcher: function( editor, matcher ) {
+			editor.plugins.clipboard._supportedFileMatchers.push( matcher );
+		},
 		/**
 		 * It returns `true` if the environment allows to set the data on copy or cut manually. This value is `false` in:
 		 * * Internet Explorer &mdash; because this browser shows the security dialog window when the script tries to set clipboard data.