+++ tests/plugins/clipboard/dropimage.js
@@ -196,7 +196,7 @@
 	bender.test( tests );
 
 	function prepareNotificationRegex( notification ) {
-		var formatsGroup = '[a-z,\\s]+',
+		var formatsGroup = '<em>[a-z/]+<\/em>',
 			regexp = '^' + notification.replace( /\$\{formats\}/g, formatsGroup ) + '$';
 
 		return new RegExp( regexp, 'gi' );