+++ tests/plugins/clipboard/notifications.js
@@ -1,0 +1,175 @@
+/* bender-tags: editor */
+/* bender-ckeditor-plugins: clipboard */
+/* bender-include: _helpers/pasting.js */
+/* bender-include: %BASE_PATH%/plugins/imagebase/features/_helpers/tools.js */
+/* global imageBaseFeaturesTools */
+
+( function() {
+	'use strict';
+
+	var config = {
+		allowedContent: true,
+		language: 'en'
+	};
+
+	bender.editors = {
+		classic: {
+			config: config
+		},
+		inline: {
+			creator: 'inline',
+			config: config
+		}
+	};
+
+	var pasteFiles = imageBaseFeaturesTools.pasteFiles,
+		tests = {
+		setUp: function() {
+			if ( !CKEDITOR.plugins.clipboard.isFileApiSupported ) {
+				assert.ignore();
+			}
+		},
+
+		'test showing notification for unsupported file types should emphasize file type': function( editor ) {
+			var notificationSpy = sinon.spy( editor, 'showNotification' ),
+				expectedMsgRegex = 'Files in <em>image/webp</em> formats are not supported',
+				file = [ { name: 'test.webp', type: 'image/webp' } ];
+
+			pasteFiles( editor, file );
+
+			resume( function() {
+				notificationSpy.restore();
+
+				assert.areSame( 1, notificationSpy.callCount, 'Notification should be called once' );
+				assert.isMatching( expectedMsgRegex, notificationSpy.getCall( 0 ).args[ 0 ],
+					'The notification has incorrect message' );
+				assert.areSame( 'info', notificationSpy.getCall( 0 ).args[ 1 ],
+					'The notification type is incorrect' );
+			}, 50 );
+
+			wait();
+		},
+
+		'test showing proper notification for file without type': function( editor ) {
+			var notificationSpy = sinon.spy( editor, 'showNotification' ),
+				expectedMsgRegex = 'File is not supported.',
+				file = [ { name: 'test1.xyz', type: '' } ];
+
+			pasteFiles( editor, file );
+
+			resume( function() {
+				notificationSpy.restore();
+
+				assert.areSame( 1, notificationSpy.callCount, 'Notification should be called once' );
+				assert.isMatching( expectedMsgRegex, notificationSpy.getCall( 0 ).args[ 0 ],
+					'The notification has incorrect message' );
+				assert.areSame( 'info', notificationSpy.getCall( 0 ).args[ 1 ],
+					'The notification type is incorrect' );
+			}, 50 );
+
+			wait();
+		},
+
+		'test notification should not be displayed when paste event do not contain files': function( editor ) {
+			var notificationSpy = sinon.spy( editor, 'showNotification' ),
+				file = [];
+
+			pasteFiles( editor, file );
+
+			resume( function() {
+				notificationSpy.restore();
+				assert.areSame( 0, notificationSpy.callCount, 'Notification should not be be called' );
+			}, 50 );
+
+			wait();
+		},
+
+		'test notification should not be displayed when paste event contains also dataValue': function( editor ) {
+			var notificationSpy = sinon.spy( editor, 'showNotification' ),
+				file = [ { name: 'test.webp', type: 'image/webp' } ],
+				dataValue = '<p>I should take precedence over files!</p>';
+
+			pasteFiles( editor, file, dataValue );
+
+			resume( function() {
+				notificationSpy.restore();
+				assert.areSame( 0, notificationSpy.callCount, 'Notification should not be be called' );
+			}, 50 );
+
+			wait();
+		},
+
+		'test showing notification for unsupported file types': function( editor ) {
+			var notificationSpy = sinon.spy( editor, 'showNotification' ),
+				expectedMsgRegex = 'Files in <em>image/cke, image/cks</em> formats are not supported',
+				files = [
+					{ name: 'test1.cke', type: 'image/cke' },
+					{ name: 'test2.cks', type: 'image/cks' }
+				];
+
+			pasteFiles( editor, files );
+
+			resume( function() {
+				notificationSpy.restore();
+
+				assert.areSame( 1, notificationSpy.callCount, 'Notification should be called once' );
+				assert.isMatching( expectedMsgRegex, notificationSpy.getCall( 0 ).args[ 0 ],
+					'The notification has incorrect message' );
+				assert.areSame( 'info', notificationSpy.getCall( 0 ).args[ 1 ],
+					'The notification type is incorrect' );
+			}, 50 );
+
+			wait();
+		},
+
+		'test showing notification for unsupported file types should not contain repeated types': function( editor ) {
+			var notificationSpy = sinon.spy( editor, 'showNotification' ),
+				expectedMsgRegex = 'Files in <em>image/webp</em> formats are not supported',
+				files = [
+					{ name: 'test1.webp', type: 'image/webp' },
+					{ name: 'test2.webp', type: 'image/webp' }
+				];
+
+			pasteFiles( editor, files );
+
+			resume( function() {
+				notificationSpy.restore();
+
+				assert.areSame( 1, notificationSpy.callCount, 'Notification should be called once' );
+				assert.isMatching( expectedMsgRegex, notificationSpy.getCall( 0 ).args[ 0 ],
+					'The notification has incorrect message' );
+				assert.areSame( 'info', notificationSpy.getCall( 0 ).args[ 1 ],
+					'The notification type is incorrect' );
+			}, 50 );
+
+			wait();
+		},
+
+		'test notification should contain only information about unsupported file types': function( editor ) {
+			var notificationSpy = sinon.spy( editor, 'showNotification' ),
+				expectedMsgRegex = 'Files in <em>application/pdf</em> formats are not supported',
+				files = [
+					{ name: 'supported.png', type: 'image/png' },
+					{ name: 'unsupported.pdf', type: 'application/pdf' }
+				];
+
+			pasteFiles( editor, files );
+
+			resume( function() {
+				notificationSpy.restore();
+
+				assert.areSame( 1, notificationSpy.callCount, 'Notification should be called once' );
+				assert.isMatching( expectedMsgRegex, notificationSpy.getCall( 0 ).args[ 0 ],
+					'The notification has incorrect message' );
+				assert.areSame( 'info', notificationSpy.getCall( 0 ).args[ 1 ],
+					'The notification type is incorrect' );
+			}, 50 );
+
+			wait();
+		}
+	};
+
+	tests = bender.tools.createTestsForEditors( CKEDITOR.tools.object.keys( bender.editors ), tests );
+
+	bender.test( tests );
+} )();